feat: restructuring the code
